Remove exclusive workspace files and move shareable files to api
