+++ src/api/clients/IOClients/infra/index.ts
@@ -4,3 +4,4 @@ export * from './Registry'
 export * from './Router'
 export * from './Runtime'
 export * from './Workspaces'
+export * from './VBase'
