+++ src/api/index.ts
@@ -11,5 +11,6 @@ export * from './locator'
 export * from './modules'
 export * from './table'
 export * from './verbose'
-export { BatchStream } from './typings'
+export * from './lib'
+export { BatchStream, WorkspaceResponse } from './typings'
 export { default as logger } from './logger'
