+++ src/api/modules/index.ts
@@ -3,3 +3,5 @@ export * from './build'
 export * from './prompts'
 export * from './utils'
 export * from './sponsor'
+export * from './workspace'
+export { default as authUrl } from './url'
