+++ src/api/modules/workspace/create.ts
@@ -1,13 +1,13 @@
 import { Workspaces } from '@vtex/api'
 import chalk from 'chalk'
-import { createFlowIssueError } from '../../api/error/utils'
-import { ErrorReport } from '../../api/error/ErrorReport'
-import { Builder } from '../../api/clients/IOClients/apps/Builder'
-import { createWorkspacesClient } from '../../api/clients/IOClients/infra/Workspaces'
-import { SessionManager } from '../../api/session/SessionManager'
-import { WorkspaceCreator } from '../../api/session/WorkspaceCreator'
-import log from '../../api/logger'
-import { promptConfirm } from '../../api/modules/prompts'
+import { createFlowIssueError } from '../../error/utils'
+import { ErrorReport } from '../../error/ErrorReport'
+import { Builder } from '../../clients/IOClients/apps/Builder'
+import { createWorkspacesClient } from '../../clients/IOClients/infra/Workspaces'
+import { SessionManager } from '../../session/SessionManager'
+import { WorkspaceCreator } from '../../session/WorkspaceCreator'
+import log from '../../logger'
+import { promptConfirm } from '../prompts'
 import { ensureValidEdition } from './common/edition'
 
 const VALID_WORKSPACE = /^[a-z][a-z0-9]{0,126}[a-z0-9]$/
