+++ src/api/typings/types.ts
@@ -3,3 +3,9 @@ export interface BatchStream {
   content: NodeJS.ReadableStream
   byteSize: number
 }
+
+export interface WorkspaceResponse {
+  name: string
+  weight: number
+  production: boolean
+}
