+++ src/modules/auth/switch.ts
@@ -4,7 +4,7 @@ import { createFlowIssueError } from '../../api/error/utils'
 import { SessionManager } from '../../api/session/SessionManager'
 import log from '../../api/logger'
 import { promptConfirm } from '../../api/modules/prompts'
-import { handleErrorCreatingWorkspace, workspaceCreator } from '../workspace/create'
+import { handleErrorCreatingWorkspace, workspaceCreator } from '../../api/modules/workspace/create'
 import welcome from './welcome'
 
 interface SwitchOptions {
